chat_pdf/requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
arxiv==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
AutoRAG==0.2.2
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bert-score==0.3.9
bitsandbytes==0.43.1
bleach==6.1.0
BLEURT==0.0.2
blinker==1.8.2
bokeh==3.4.1
boto3==1.34.117
botocore==1.34.117
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.3
cohere==5.5.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cryptography==42.0.7
cycler==0.12.1
Cython==3.0.10
dask==2024.5.2
dask-expr==1.1.2
dataclasses-json==0.6.6
datasets==2.10.1
debugpy==1.6.7
decorator==5.1.1
deepeval==0.21.43
Deprecated==1.2.14
dill==0.3.6
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docx2txt==0.8
email_validator==2.1.1
entrypoints==0.4
evaluate==0.4.2
exceptiongroup==1.2.0
execnet==2.1.1
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
fastparquet==2024.5.0
feedparser==6.0.10
filelock==3.14.0
FlagEmbedding==1.2.10
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
griffe==0.45.1
grpcio==1.64.0
guidance==0.1.15
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
interegular==0.3.3
ipykernel==6.29.3
ipython==8.24.0
ipywidgets==8.1.2
ipywidgets_bokeh==1.6.0
jedi==0.19.1
Jinja2==3.1.4
jiwer==2.5.1
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
jq==1.7.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
keras==3.3.3
kiwipiepy==0.17.1
kiwipiepy_model==0.17.0
kiwisolver==1.4.5
konlpy==0.6.0
kubernetes==29.0.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langdetect==1.0.9
langsmith==0.1.60
lark==1.1.9
Levenshtein==0.20.2
libclang==18.1.1
linkify-it-py==2.0.3
llama-index==0.10.42
llama-index-agent-openai==0.2.6
llama-index-cli==0.1.12
llama-index-core==0.10.42
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-huggingface==0.2.2
llama-index-llms-openai==0.1.21
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-retrievers-bm25==0.1.3
llama-index-vector-stores-chroma==0.1.8
llama-parse==0.4.4
llamaindex-py-client==0.1.19
llmlingua==0.2.2
llvmlite==0.42.0
lm-format-enforcer==0.10.1
locket==1.0.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
minijinja==2.0.1
mixpanel==4.10.1
ml-dtypes==0.3.2
mlflow-skinny==2.13.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.14
mypy-extensions==1.0.0
namex==0.0.8
nest_asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.30.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
orjson==3.10.3
outlines==0.0.34
overrides==7.7.0
packaging==23.2
pandas==2.2.2
panel==1.4.4
param==2.1.0
parso==0.8.4
partd==1.4.2
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
pluggy==1.5.0
poc-rag==0.1
portalocker==2.8.2
posthog==3.5.0
prometheus_client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
prompt-toolkit==3.0.42
protobuf==4.25.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydot==2.0.0
pyformlang==1.0.10
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pynndescent==0.5.12
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
pytest==8.2.1
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyviz_comms==3.0.2
PyYAML==6.0.1
pyzmq==25.1.2
ragas==0.1.8
rank-bm25==0.2.2
rapidfuzz==2.15.2
ray==2.23.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.7.1
rouge==1.0.1
rouge-metric==1.0.1
rouge_score==0.1.2
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.11.4
seaborn==0.13.2
sentence-transformers==2.7.0
sentencepiece==0.2.0
sentry-sdk==2.2.0
setuptools==69.5.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.2
starlette==0.37.2
streamlit==1.35.0
striprtf==0.0.26
swifter==1.4.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
text-generation==0.7.0
tf-slim==1.1.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tokenlog==0.0.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
triton==2.3.0
typer==0.12.3
types-requests==2.32.0.20240523
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vllm==0.4.3
vllm-flash-attn==2.5.8.post2
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xformers==0.0.26.post1
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.18.2
zstandard==0.22.0
